Modified Files: 	COPYING configure.ac rdiff source/Makefile.am 	source/org/metastatic/rsync/BrokenMD4.java 	source/org/metastatic/rsync/Checksum32.java 	source/org/metastatic/rsync/ChecksumPair.java 	source/org/metastatic/rsync/Configuration.java 	source/org/metastatic/rsync/DataBlock.java 	source/org/metastatic/rsync/Delta.java 	source/org/metastatic/rsync/Generator.java 	source/org/metastatic/rsync/GeneratorEvent.java 	source/org/metastatic/rsync/GeneratorListener.java 	source/org/metastatic/rsync/GeneratorStream.java 	source/org/metastatic/rsync/JarsyncProvider.java 	source/org/metastatic/rsync/MD4.java 	source/org/metastatic/rsync/Matcher.java 	source/org/metastatic/rsync/MatcherEvent.java 	source/org/metastatic/rsync/MatcherListener.java 	source/org/metastatic/rsync/MatcherStream.java 	source/org/metastatic/rsync/Offsets.java 	source/org/metastatic/rsync/ParameterException.java 	source/org/metastatic/rsync/ParameterListener.java 	source/org/metastatic/rsync/Parameters.java 	source/org/metastatic/rsync/Rdiff.java 	source/org/metastatic/rsync/Rebuilder.java 	source/org/metastatic/rsync/RebuilderEvent.java 	source/org/metastatic/rsync/RebuilderListener.java 	source/org/metastatic/rsync/RebuilderStream.java 	source/org/metastatic/rsync/RollingChecksum.java 	source/org/metastatic/rsync/TwoKeyMap.java 	source/org/metastatic/rsync/Util.java 	source/org/metastatic/rsync/v2/BlockingDaemon.java 	source/org/metastatic/rsync/v2/Constants.java 	source/org/metastatic/rsync/v2/FileInfo.java 	source/org/metastatic/rsync/v2/FileList.java 	source/org/metastatic/rsync/v2/NonblockingReceiver.java 	source/org/metastatic/rsync/v2/NonblockingSender.java 	source/org/metastatic/rsync/v2/Options.java 	source/org/metastatic/rsync/v2/Protocol.java 	source/org/metastatic/rsync/v2/Receiver.java 	source/org/metastatic/rsync/v2/Sender.java 	test/gnu/testlet/org/metastatic/rsync/SimpleTest.java 	test/gnu/testlet/org/metastatic/rsync/SimpleTest2.java 	test/gnu/testlet/org/metastatic/rsync/TestOfBrokenMD4.java 	test/gnu/testlet/org/metastatic/rsync/TestOfChecksum32.java 	test/gnu/testlet/org/metastatic/rsync/TestOfMD4.java
	Fuck it. Lots of little changes, too lazy.


git-svn-id: https://jarsync.svn.sourceforge.net/svnroot/jarsync@160 388fd5d1-5327-0410-80e4-b6d47a689578

+++ trunk/jarsync/source/org/metastatic/rsync/Checksum32.java
@@ -59,21 +59,21 @@ public class Checksum32 implements RollingChecksum, Cloneable, java.io.Serializa
    // Constants and variables.
    // -----------------------------------------------------------------
 
-   private final short char_offset;
+   private final int char_offset;
 
    /**
     * The first half of the checksum.
     *
     * @since 1.1
     */
-   protected short a;
+   protected int a;
 
    /**
     * The second half of the checksum.
     *
     * @since 1.1
     */
-   protected short b;
+   protected int b;
 
    /**
     * The place from whence the current checksum has been computed.
@@ -119,14 +119,14 @@ public class Checksum32 implements RollingChecksum, Cloneable, java.io.Serializa
     * affects the output of this checksum; rsync uses a char offset of
     * 0, librsync 31.
     */
-   public Checksum32(short char_offset) {
+   public Checksum32(int char_offset) {
       this.char_offset = char_offset;
       a = b = 0;
       k = 0;
    }
 
    public Checksum32() {
-      this((short) 0);
+      this(0);
    }
 
    private Checksum32(Checksum32 that) {
@@ -151,7 +151,7 @@ public class Checksum32 implements RollingChecksum, Cloneable, java.io.Serializa
     * @since 1.1
     */
    public int getValue() {
-      return (a&0xffff) | (b << 16);
+      return (a & 0xffff) | (b << 16);
    }
 
    /**
