+++ trunk/jarsync/test/gnu/testlet/org/metastatic/rsync/TestOfChecksum32.java
@@ -1,34 +1,34 @@
 /* vim:set softtabstop=3 shiftwidth=3 tabstop=3 expandtab tw=72:
    $Id $
-  
+
    TestOfChecksum32: test of the rolling checksum.
    Copyright (C) 2003  Casey Marshall <rsdio@metastatic.org>
-  
+
    This file is a part of Jarsync
-  
+
    Jarsync is free software; you can redistribute it and/or modify it
    under the terms of the GNU General Public License as published by the
    Free Software Foundation; either version 2 of the License, or (at
    your option) any later version.
-  
+
    Jarsync is distributed in the hope that it will be useful, but
    WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
    General Public License for more details.
-  
+
    You should have received a copy of the GNU General Public License
    along with Jarsync; if not, write to the
-  
+
       Free Software Foundation, Inc.,
       59 Temple Place, Suite 330,
       Boston, MA  02111-1307
       USA
-  
+
    Linking Jarsync statically or dynamically with other modules is
    making a combined work based on Jarsync.  Thus, the terms and
    conditions of the GNU General Public License cover the whole
    combination.
-  
+
    As a special exception, the copyright holders of Jarsync give you
    permission to link Jarsync with independent modules to produce an
    executable, regardless of the license terms of these independent
@@ -62,7 +62,7 @@ public class TestOfChecksum32 implements Testlet {
 
    public void test(TestHarness harness) {
       try {
-         Checksum32 c = new Checksum32((short) 0);
+         Checksum32 c = new Checksum32(0);
          c.check("a".getBytes(), 0, 1);
          harness.check(c.getValue() == 0x610061, "TestA");
       } catch (Exception x) {
@@ -71,7 +71,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 0);
+         Checksum32 c = new Checksum32(0);
          c.check("abc".getBytes(), 0, 3);
          harness.check(c.getValue() == 0x24a0126, "TestABC");
       } catch (Exception x) {
@@ -80,7 +80,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 0);
+         Checksum32 c = new Checksum32(0);
          c.check("abcdefghijklmnopqrstuvwxyz".getBytes(), 0, 26);
          harness.check(c.getValue() == 0x906c0b1f, "TestAlphabet");
       } catch (Exception x) {
@@ -89,7 +89,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 0);
+         Checksum32 c = new Checksum32(0);
          c.check("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890".getBytes(), 0, 62);
          harness.check(c.getValue() == 0xdf2c150b, "TestASCIISubset");
       } catch (Exception x) {
@@ -98,7 +98,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 0);
+         Checksum32 c = new Checksum32(0);
          c.check("Adler 32".getBytes(), 0, 8);
          harness.check(c.getValue() == 0xc05026d, "TestAdler32");
       } catch (Exception x) {
@@ -107,7 +107,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 0);
+         Checksum32 c = new Checksum32(0);
          c.check("01234567890123456789012345678901234567890123456789012345678901234567890123456789".getBytes(), 0, 80);
          harness.check(c.getValue() == 0x95e01068, "TestEightyNumerics");
       } catch (Exception x) {
@@ -116,8 +116,8 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 0);
-         c.check(new byte[] { 
+         Checksum32 c = new Checksum32(0);
+         c.check(new byte[] {
             (byte) 0xDE, (byte) 0xAD, (byte) 0xBE, (byte) 0xEF
          }, 0, 4);
          harness.check(c.getValue() == 0xfdeaff38, "TestDEADBEEF");
@@ -127,8 +127,8 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 0);
-         c.check(new byte[] { 
+         Checksum32 c = new Checksum32(0);
+         c.check(new byte[] {
             (byte) 0xCA, (byte) 0xFE, (byte) 0xBA, (byte) 0xBE
          }, 0, 4);
          harness.check(c.getValue() == 0xfe54ff40, "TestCAFEBABE");
@@ -138,7 +138,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 0);
+         Checksum32 c = new Checksum32(0);
          byte[] buf = new byte[256];
          for (int i = 0; i < 256; i++)
             buf[i] = (byte) i;
@@ -153,8 +153,8 @@ public class TestOfChecksum32 implements Testlet {
          Random r = new Random();
          harness.checkPoint("testRoll");
          for (int i = 0; i < 10; i++) {
-            Checksum32 c1 = new Checksum32((short) 0);
-            Checksum32 c2 = new Checksum32((short) 0);
+            Checksum32 c1 = new Checksum32(0);
+            Checksum32 c2 = new Checksum32(0);
             byte[] buf1 = new byte[200 + r.nextInt(1200)];
             byte[] buf2 = new byte[buf1.length];
             r.nextBytes(buf1);
@@ -175,7 +175,7 @@ public class TestOfChecksum32 implements Testlet {
       // CHAR_OFFSET = 31 ---------------------------------------------------
 
       try {
-         Checksum32 c = new Checksum32((short) 31);
+         Checksum32 c = new Checksum32(31);
          c.check("a".getBytes(), 0, 1);
          harness.check(c.getValue() == 0x800080, "TestA");
       } catch (Exception x) {
@@ -184,7 +184,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 31);
+         Checksum32 c = new Checksum32(31);
          c.check("abc".getBytes(), 0, 3);
          harness.check(c.getValue() == 0x3040183, "TestABC");
       } catch (Exception x) {
@@ -193,7 +193,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 31);
+         Checksum32 c = new Checksum32(31);
          c.check("abcdefghijklmnopqrstuvwxyz".getBytes(), 0, 26);
          harness.check(c.getValue() == 0xbaed0e45, "TestAlphabet");
       } catch (Exception x) {
@@ -202,7 +202,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 31);
+         Checksum32 c = new Checksum32(31);
          c.check("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890".getBytes(), 0, 62);
          harness.check(c.getValue() == 0xcbab1c8d, "TestASCIISubset");
       } catch (Exception x) {
@@ -211,7 +211,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 31);
+         Checksum32 c = new Checksum32(31);
          c.check("Adler 32".getBytes(), 0, 8);
          harness.check(c.getValue() == 0x10610365, "TestAdler32");
       } catch (Exception x) {
@@ -220,7 +220,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 31);
+         Checksum32 c = new Checksum32(31);
          c.check("01234567890123456789012345678901234567890123456789012345678901234567890123456789".getBytes(), 0, 80);
          harness.check(c.getValue() == 0x1e381a18, "TestEightyNumerics");
       } catch (Exception x) {
@@ -229,8 +229,8 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 31);
-         c.check(new byte[] { 
+         Checksum32 c = new Checksum32(31);
+         c.check(new byte[] {
             (byte) 0xDE, (byte) 0xAD, (byte) 0xBE, (byte) 0xEF
          }, 0, 4);
          harness.check(c.getValue() == 0xff20ffb4, "TestDEADBEEF");
@@ -240,8 +240,8 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 31);
-         c.check(new byte[] { 
+         Checksum32 c = new Checksum32(31);
+         c.check(new byte[] {
             (byte) 0xCA, (byte) 0xFE, (byte) 0xBA, (byte) 0xBE
          }, 0, 4);
          harness.check(c.getValue() == 0xff8affbc, "TestCAFEBABE");
@@ -251,7 +251,7 @@ public class TestOfChecksum32 implements Testlet {
       }
 
       try {
-         Checksum32 c = new Checksum32((short) 31);
+         Checksum32 c = new Checksum32(31);
          byte[] buf = new byte[256];
          for (int i = 0; i < 256; i++)
             buf[i] = (byte) i;
@@ -266,8 +266,8 @@ public class TestOfChecksum32 implements Testlet {
          Random r = new Random();
          harness.checkPoint("testRoll");
          for (int i = 0; i < 10; i++) {
-            Checksum32 c1 = new Checksum32((short) 31);
-            Checksum32 c2 = new Checksum32((short) 31);
+            Checksum32 c1 = new Checksum32(31);
+            Checksum32 c2 = new Checksum32(31);
             byte[] buf1 = new byte[201 + r.nextInt(1200)];
             r.nextBytes(buf1);
             c1.check(buf1, 0, buf1.length-1);
